Added an initial countdown before each trial.
Before each trial begins, the screen is greyed-out for a short time (configurable in settings.json), to allow the user to observe the test before beginning. This minimizes the impact of the user's reaction time on test results, as they get a chance to familiarize themselves with the configuration before beginning. No movements/statistics are applied/calculated during this initial countdown period.