Fix. Control Manager now scales sensor reading considering activationThreshold.
The activationThreshold setting is used to ignore/eliminate cursor jitter caused by small, involuntary muscle contractions that occur when the player is at rest. Any sensor reading smaller than the activationThreshold is treated as unintentional noise and explicitly ignored. However, the remaining range of sensor readings was previously not being scaled in any waay, resulting in an abrupt spike of movement as soon as the activationThreshold was exceeded (e.g., with activationThreshold = 25, the cursor transitions from being completely motionless to ~ 1/4 it's maximum speed)

To fix this issue, the remaining range of sensor readings is now scaled appropriately using the activationThreshold setting to ensure that the full range of movement speeds is still attainable (e.g., a reading of 0.21 with activationThreshold = 0.20 now results in a movement speed of 1.25% of maximum, instead of an abrupt jump to 21% of maximum)